// src/types/materials.ts
export type LearningMaterial = {
  _id: string;
  title: string;
  age_group: string[];
  developmental_areas: string[];
  learning_objective_tags: string[];
  interest_tags: string[];
  content_type: string[];
  engagement_level: string[];
  duration: string;
  skill_level: string;
  document_link: string;
};